+++ scripts/debian/build-deb.sh
@@ -233,7 +233,7 @@ if [[ ! -d "${BUILD_DIR}" ]]; then
     fi
   fi
 
-  install-libzippp
+  install-libzippp static
   install-fmt
   install-spdlog
   install-taglib