+++ gerbera-web/test/e2e/config.spec.js
@@ -56,7 +56,7 @@ describe('Config Suite', () => {
       await homePage.showConfig('Server');
 
       let result = await homePage.editOverlayFieldValue('grb_value__server_modelNumber');
-      expect(result).to.equal('2.6.0');
+      expect(result).to.equal('3.1.0');
 
       await homePage.setEditorOverlayField('grb_value__server_modelNumber', '43');